requirements.txt
opencv-python==4.5.5.64
pyzbar==0.1.8
Pillow==9.3.0
requests==2.28.1